// src/config/generatedFolderIconManifest.ts
export type GeneratedFolderIconPair = {
  closed: string;
  open: string | null;
};

export const DEFAULT_FOLDER_ICON = {"closed":"folder.svg","open":"folder_open.svg"} as const;

export const CUSTOM_FOLDER_ICONS = {
// ...
  "public": {"closed":"folder_custom_public.svg","open":"folder_custom_public.svg"},
// ...
} as const satisfies Record<string, GeneratedFolderIconPair>;